Fix player shielding when CPU is set to shield during L frame advance
A shotgun of changes here, basically we want to make sure that the
engine and master pads are disabled on L press when frame advance is
set to L. We also want to make sure that we're only removing inputs
if the frame advance is enabled in the first place. Additionally, we
want to make sure that the CPU is shielding if it's supposed to shield.
Some of these changes might not be necessary, I was working on this
one for a while...