Add additional SDI options and make SDI function properly
Add SDI in different cardinal directions, and also a "Very High"
option that will trigger on every frame. Also fix SDI to do more than
one SDI, it will now alternate between the cardinal and a region
a bit out of the deadzone. We might want to adjust the random values
for these.