Remove snapshot functionality from import UI
The existing implementation is flawed causing data races when deleting
recordings and slows down the overall UI with the extra memory card
operations. The functionality could be restored later, but it would
require a more thorough redesign of the UI to work well. Removing for
now to prevent the delete hanging the UI makes the most sense.